Separate different parts of the app::start method into respective mods
We have several different modules now that can contain the logic that
currently lives in app::start. Move the logic to the appropriate
modules, then app::start is simplified to just coordinate calling all
the individual pieces.

Closes https://github.com/roadster-rs/roadster/issues/9
