+++ src/cli/roadster/open_api_schema.rs
